+++ main.ts
@@ -1,8 +1,9 @@
 import { Application, Context, oakCors } from "./config/deps.ts";
-import { DEV_MODE, PORT } from "./config/env.ts";
+import { PORT } from "./config/env.ts";
 import { rateLimitMiddleware } from "./middleware/ratelimitMiddleware.ts";
 import router from "./routes.ts";
 import { responseLog } from "./middleware/ResponseLog.ts";
+import { log, logError } from "./utils/utils.ts";
 
 const app = new Application();
 
@@ -14,6 +15,7 @@ app.use(async (ctx: Context, next) => {
 });
 
 app.use(oakCors());
+
 app.use(responseLog);
 app.use(rateLimitMiddleware);
 
@@ -25,15 +27,9 @@ app.use((ctx: Context) => {
   ctx.response.body = { error: "Endpoint not found" };
 });
 
-app.addEventListener(
-  "error",
-  (evt) =>
-    console.error(`[${new Date().toISOString()}] Unhandled error:`, evt.error),
-);
+app.addEventListener("error", (evt) => {
+  logError(`Unhandled error:`, evt.error);
+});
 
-DEV_MODE &&
-  console.log(
-    `[${new Date().toISOString()}] Stremio AI Addon running on port ${PORT}`,
-  );
-
+log(`Stremio AI Addon running on port ${PORT}`);
 await app.listen({ hostname: "0.0.0.0", port: PORT });